♻️ : 게임 크롤링 서비스 리팩토링 - 불필요한 서비스 제거 및 새로운 유틸리티 함수 추가, 크론잡 설정
