+++ src/modules/game/core/domain/game.entity.ts
@@ -11,7 +11,7 @@ import {
 import { Team } from 'src/modules/team/entities/team.entity';
 import { Stadium } from 'src/modules/stadium/entities/stadium.entity';
 import { RegisteredGame } from 'src/modules/registered-game/core/domain/registered-game.entity';
-import { TGameStatus } from 'src/modules/scheduling/crawling-game.type';
+import { TGameStatus } from 'src/modules/game/application/util/crawling-game.type';
 
 @Entity()
 @Unique(['date', 'time', 'stadium'])
