--- src/modules/game/infra/crawling/game-crawling.module.ts
@@ -1,9 +1,0 @@
-import { Module } from '@nestjs/common';
-import { GameScheduleCrawlingService } from './game-schedule.crawling.service';
-import { GameScoreCrawlingService } from './game-score.crawling.service';
-
-@Module({
-  providers: [GameScheduleCrawlingService, GameScoreCrawlingService],
-  exports: [GameScheduleCrawlingService, GameScoreCrawlingService],
-})
-export class GameCrawlingModule {}
