--- src/modules/game/util/gamel-list-group-by-team.util.ts
@@ -1,17 +1,0 @@
-import { GameDto } from '../dto/game.dto';
-
-export function groupGamesByTeam(games: GameDto[]): Record<string, GameDto[]> {
-  const map: Record<string, GameDto[]> = {};
-
-  for (const game of games) {
-    const key = game.id.slice(8, -1);
-
-    if (!map[key]) {
-      map[key] = [];
-    }
-
-    map[key].push(game);
-  }
-
-  return map;
-}
